Compact FMData class down some

+++ AngelLoader/Common/GameSupport.cs
@@ -20,7 +20,8 @@ public static partial class GameSupport
 
     // This is flags so we can combine its values for filtering by multiple games.
     [Flags, FenGenGameEnum(gameIndexEnumName: "GameIndex")]
-    public enum Game : uint
+    // @vNext: Generate type for GameIndex
+    public enum Game : byte
     {
         [FenGenIgnore]
         Null = 0,
